+++ neo-cli/CLI/MainService.NEP17.cs
@@ -22,10 +22,10 @@ namespace Neo.CLI
         /// <param name="from">From</param>
         /// <param name="signersAccounts">Signer's accounts</param>
         [ConsoleCommand("transfer", Category = "NEP17 Commands")]
-        private void OnTransferCommand(UInt160 tokenHash, UInt160 to, long amount, string data = null, UInt160 from = null, UInt160[] signersAccounts = null)
+        private void OnTransferCommand(UInt160 tokenHash, UInt160 to, decimal amount, string data = null, UInt160 from = null, UInt160[] signersAccounts = null)
         {
             var asset = new AssetDescriptor(tokenHash);
-            var value = new BigDecimal((decimal)amount, asset.Decimals);
+            var value = new BigDecimal(amount, asset.Decimals);
 
             if (NoWallet()) return;
